Switch to submodules

+++ DynmapBukkitBase/src/main/java/org/dynmap/bukkit/BukkitVersionHelperGlowstone.java
@@ -445,7 +445,7 @@ public class BukkitVersionHelperGlowstone extends BukkitVersionHelper {
 
     @Override
     public int getHealth(Player p) {
-        return p.getHealth();
+        return (int) p.getHealth();
     }
 
 }
